pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantproto"
version = "0.1.0"
description = "Reproducible quant research engine"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "hmmlearn>=0.3",
    "matplotlib>=3.7",
    "fastapi>=0.100",
    "fastmcp>=0.1",
    "redis>=5.0",
    "PyJWT>=2.8",
    "uvicorn>=0.23",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
include = ["quantproto*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
